Update vector db onboarding code examples
