+++ x-pack/solutions/vectordb/packages/kbn-vectordb-onboarding/src/onboarding/components/api_step.test.tsx
@@ -11,8 +11,8 @@ import { EuiCopy } from '@elastic/eui';
 import { KibanaContextProvider } from '@kbn/kibana-react-plugin/public';
 import { TryInConsoleButton } from '@kbn/try-in-console';
 import { ApiStep, type ApiStepTab } from './api_step';
-import { LANGUAGES } from './languages';
-import { API_KEY_PLACEHOLDER, URL_PLACEHOLDER } from './console_snippets';
+import { LANGUAGES } from '../constants/languages';
+import { API_KEY_PLACEHOLDER, URL_PLACEHOLDER } from '../constants/console_snippets';
 import { useOnboardingCredentials } from '../../hooks/use_onboarding_credentials';
 import type { OnboardingServices } from '../../services';
