+++ x-pack/solutions/vectordb/packages/kbn-vectordb-onboarding/src/onboarding/constants/console_snippets.test.ts
