+++ x-pack/solutions/vectordb/packages/kbn-vectordb-onboarding/src/onboarding/constants/languages.ts
